if max validators reached then break out of iteration over candidates, improve comments

+++ pallets/stake/src/mock.rs
@@ -182,7 +182,7 @@ impl pallet_authorship::Config for Test {
 	type EventHandler = Module<Test>;
 }
 parameter_types! {
-	pub const MaxValidators: usize = 5;
+	pub const MaxValidators: u32 = 5;
 	pub const MaxNominatorsPerValidator: usize = 10;
 	pub const MinNominatorsPerValidator: usize = 0;
 	pub const MinCandidateBond: u128 = 10;
